requirements.txt
Pillow
accelerate
addict
aiofiles
aiohttp
aiosignal
annotated-types
anyio
appdirs
async-timeout
attrs
beartype
bitsandbytes
blis
clip
datasets
diffusers
einops
gradio
huggingface-hub
jsonlines
kornia
mediapipe
numpy
omegaconf
openai-clip
opencv-python
packaging
pycocotools
pytorch-lightning
pyyaml
scikit-image
sentencepiece
spacy
taming-transformers
tensorboard
termcolor
timm
tokenizers
torch==2.0.1
torchtyping
torchvision==0.15.2
tqdm
transformers
tslearn
xformers
yapf
